test: remove usage of remote in shell.openExternal test
